Cloud Functions/ingest_webshop/requirements.txt
google-cloud-storage
datetime
pymysql